+++ models/comment.js
@@ -1,15 +1,25 @@
-'use strict'
+'use strict';
+const {
+  Model
+} = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
-  const comment = sequelize.define('comment', {
+  class comment extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      // define association here
+    }
+  };
+  comment.init({
     title: DataTypes.STRING,
     content: DataTypes.TEXT,
     articleId: DataTypes.INTEGER
-  }, {})
-
-  comment.associate = function(models) {
-    // associations can be defined here
-    models.comment.belongsTo(models.article)
-  }
-  
-  return comment
-}
+  }, {
+    sequelize,
+    modelName: 'comment',
+  });
+  return comment;
+};
