+++ models/commentsss.js
@@ -1,0 +1,15 @@
+'use strict'
+module.exports = (sequelize, DataTypes) => {
+  const comment = sequelize.define('comment', {
+    title: DataTypes.STRING,
+    content: DataTypes.TEXT,
+    articleId: DataTypes.INTEGER
+  }, {})
+
+  comment.associate = function(models) {
+    // associations can be defined here
+    models.comment.belongsTo(models.article)
+  }
+  
+  return comment
+}
